testing backend and assorted refactoring
