+++ snuba/utils/metrics/types.py
@@ -1,0 +1,3 @@
+from typing import Mapping
+
+Tags = Mapping[str, str]
